NT project compile done.

+++ ctl/component/intrinsic/advance/src/ctl_adv_imc.c
@@ -8,17 +8,17 @@ int ctl_init_imc(ctl_imc_controller_t* imc, const ctl_imc_init_t* init)
 {
     // 1. 防呆与除零保护
     gmp_base_assert(init->f_ctrl > 0.0);
-    gmp_base_assert(fabs(init->K_p) > 1e-9); // Kp 不能为 0，因为后续要作除数
+    gmp_base_assert(fabsf(init->K_p) > 1e-9f); // Kp 不能为 0，因为后续要作除数
 
     // 2. 纯物理参数域计算，严禁出现 ctrl_gt 强转！
     parameter_gt Ts = 1.0 / init->f_ctrl;
 
     // --- Discretize Plant Model (ZOH) ---
-    parameter_gt a_p_d_f = exp(-(Ts / init->tau_p));
+    parameter_gt a_p_d_f = expf(-(Ts / init->tau_p));
     parameter_gt b_p_d_f = init->K_p * (1.0 - a_p_d_f);
 
     // --- Calculate Dead Time ---
-    imc->dead_time_samples = (uint16_t)round(init->theta_p / Ts);
+    imc->dead_time_samples = (uint16_t)roundf(init->theta_p / Ts);
     if (imc->dead_time_samples >= IMC_MAX_DEAD_TIME_SAMPLES)
     {
         return -1; // Error: Dead time exceeds buffer size
